+++ CppNagyHF/RandomModel.cpp
@@ -31,7 +31,7 @@ namespace model {
 		// find max value
 		int choice = 0;
 		double maxValue = result[0];
-		for (size_t i = 1; i < std::min<int>(result.size(), 4); i++) {
+		for (int i = 1; i < std::min<int>((int)result.size(), 4); i++) {
 			if (result[i] > maxValue) {
 				maxValue = result[i];
 				choice = i;
